General cleanup and one major bugfix
The bugfix is that non-root names were not collected before we actually
started trying to use them to distinguish roots.  I was able to
collect non-root names entirely in SQL by looking for fragments that
were references, and that belonged to code sections and found the name
id by comparing with that fragment's data.